shelf_ai/requirements.txt
# Core ML dependencies
ultralytics>=8.0.0          # YOLOv8 (includes PyTorch)

# Dashboard
streamlit>=1.28.0

# Image processing
opencv-python>=4.8.0
Pillow>=9.0.0

# Config / data
pyyaml>=6.0
numpy>=1.24.0

# Testing
pytest>=7.4.0
